requirements.txt
# SpoonOS for agentic architecture
spoon-ai-sdk

# Optional: Extended toolkits
spoon-toolkits

# Environment variable management
python-dotenv

# RAG and Vector Database
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0


# Google Generative AI
google-generativeai
requests
